Removed filewriter.go and all associated references
