+++ data/dataaccessor.go
@@ -16,7 +16,7 @@ func NewAccessor(access DataAccess) DataAccessor {
 }
 
 /*
-DataAccess represents an interface for accessing and managing a data storage system (e.g. a filesystem or database).
+DataAccess represents an interface for accessing and managing a data storage system (e.g. a database).
 Implementations of DataAccess should ensure that the objects passed to them are properly serialized under, and can be
 retrieved by providing, the specified key string.
 */
